Parse Info.plist instead of identifier.
The identifier that is used in stack traces corresponds with the CFBundleIdentifier instead of the actual executable.  Instead of just stripping off the last portion of the identifier and using it as the executable, this patch will parse the Info.plist file, match the CFBundleIdentifier, and use the CFBundleExecutable instead.